Fix item loss with drop transfer and full inv (#14)
When a drop transfer chest is full, the module will attempt to return
excess items to the player. However, this doesn't account for when
the player's inventory also becomes full (e.g. because of other items
in the ground).

This situation can arise if a player holds an item under their cursor,
and suddenly their inventory becomes full. The item under the cursor
becomes lost, as LWC destroys it.